pyproject.toml
[tool.poetry]
name = "mader"
version = "0.1.0"
description = ""
authors = ["Thalison <thalissfernandes99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.0"}
pydantic-settings = "^2.4.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
aiosqlite = "^0.20.0"
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
coverage = "^7.6.0"


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']


[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'


[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev mader/app.py'
pre_test = 'task lint'
test = 'pytest -s --cov=mader -vv'
post_test = 'coverage html'


[tool.poetry.group.dev.dependencies]
ignr = "^2.2"
taskipy = "^1.13.0"
ruff = "^0.5.6"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.8"
alembic = "^1.13.2"
gevent = "^24.2.1"
freezegun = "^1.5.1"
factory-boy = "^3.3.0"

[tool.coverage.run]
concurrency = ["greenlet", "thread"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
